Massive resync of my tree and freebios-cvs including. - Motherboard identification in the linuxbios table. - Better SMP startup code. - Support for the tyan s2462, the supermicro p4dc6, the supermicro p4dc6+, and the beginnings of supermicro p4dpr support - Remove of no longer needed USE_AMD_NDA option

+++ src/northbridge/amd/amd76x/northbridge.c
@@ -4,7 +4,7 @@
 
 unsigned long sizeram(void)
 {
-	unsigned long size;
+	u32 size;
 	/* Use the PCI top memory register */
 	pcibios_read_config_dword(0, 0, 0x9c, &size);
 	/* Convert size in bytes to size in K */
